car-damage-main/requirements.txt
flask
google-generativeai
python-dotenv
Pillow
werkzeug
gunicorn
psycopg2-binary
